Pull out pubkey functionality into @cosmjs/amino
