+++ packages/amino/.eslintignore
@@ -1,0 +1,8 @@
+node_modules/
+
+build/
+custom_types/
+dist/
+docs/
+generated/
+types/
