+++ packages/amino/.nycrc.yml
@@ -1,0 +1,1 @@
+../../.nycrc.yml
