+++ packages/amino/README.md
@@ -1,0 +1,12 @@
+# @cosmjs/amino
+
+[![npm version](https://img.shields.io/npm/v/@cosmjs/amino.svg)](https://www.npmjs.com/package/@cosmjs/amino)
+
+Helpers for Amino based signing which are shared between @cosmjs/launchpad and
+@cosmjs/stargate.
+
+## License
+
+This package is part of the cosmjs repository, licensed under the Apache License
+2.0 (see [NOTICE](https://github.com/cosmos/cosmjs/blob/main/NOTICE) and
+[LICENSE](https://github.com/cosmos/cosmjs/blob/main/LICENSE)).
