+++ packages/amino/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  decodeAminoPubkey,
+  decodeBech32Pubkey,
+  encodeAminoPubkey,
+  encodeBech32Pubkey,
+  encodeSecp256k1Pubkey,
+} from "./encoding";
+export { PubKey, pubkeyType } from "./pubkeys";
