+++ packages/launchpad/package.json
@@ -39,6 +39,7 @@
     "pack-web": "yarn build-or-skip && webpack --mode development --config webpack.web.config.js"
   },
   "dependencies": {
+    "@cosmjs/amino": "^0.25.0-alpha.0",
     "@cosmjs/crypto": "^0.25.0-alpha.0",
     "@cosmjs/encoding": "^0.25.0-alpha.0",
     "@cosmjs/math": "^0.25.0-alpha.0",
