+++ packages/launchpad/src/types.ts
@@ -1,4 +1,6 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import { PubKey } from "@cosmjs/amino";
+
 import { Coin } from "./coins";
 
 export interface StdFee {
@@ -10,22 +12,3 @@ export interface StdSignature {
   readonly pub_key: PubKey;
   readonly signature: string;
 }
-
-export interface PubKey {
-  // type is one of the strings defined in pubkeyType
-  // I don't use a string literal union here as that makes trouble with json test data:
-  // https://github.com/cosmos/cosmjs/pull/44#pullrequestreview-353280504
-  readonly type: string;
-  // Value field is base64-encoded in all cases
-  // Note: if type is Secp256k1, this must contain a COMPRESSED pubkey - to encode from bcp/keycontrol land, you must compress it first
-  readonly value: string;
-}
-
-export const pubkeyType = {
-  /** @see https://github.com/tendermint/tendermint/blob/v0.33.0/crypto/ed25519/ed25519.go#L22 */
-  secp256k1: "tendermint/PubKeySecp256k1" as const,
-  /** @see https://github.com/tendermint/tendermint/blob/v0.33.0/crypto/secp256k1/secp256k1.go#L23 */
-  ed25519: "tendermint/PubKeyEd25519" as const,
-  /** @see https://github.com/tendermint/tendermint/blob/v0.33.0/crypto/sr25519/codec.go#L12 */
-  sr25519: "tendermint/PubKeySr25519" as const,
-};
