+++ packages/stargate/package.json
@@ -43,6 +43,7 @@
   },
   "dependencies": {
     "@confio/ics23": "^0.6.3",
+    "@cosmjs/amino": "^0.25.0-alpha.0",
     "@cosmjs/encoding": "^0.25.0-alpha.0",
     "@cosmjs/launchpad": "^0.25.0-alpha.0",
     "@cosmjs/math": "^0.25.0-alpha.0",
